Create Data scraping of Amazon.in website using Python - BeautifulSoup
The explosion of the internet has been a boon for data enthusiasts. The variety and quantity of data that is available today through the internet is like a treasure trove of secrets and mysteries waiting to be solved. For example, you are planning to travel – how about scraping a few travel recommendation sites, pull out comments about various do to things and see which property is getting a lot of positive responses from the users! The list of use cases is endless.

Yet, there is no fixed methodology to extract such data and much of it is unstructured and full of noise.
Such conditions make web scraping a necessary technique for a data scientist’s toolkit. As it is rightfully said, "Any content that can be viewed on a webpage can be scraped. Period."

In this example, we will scarpe Amazon.in website to get a list of mobile phones with other related information. We will use Python for programming and BeautifulSoup library to scrape the data.